+++ CotEditor/Sources/CECharacterPopoverController.m
@@ -26,26 +26,18 @@
  */
 
 #import "CECharacterPopoverController.h"
-
-
-// variation selectors
-static const unichar  kTextSequenceChar = 0xFE0E;
-static const unichar kEmojiSequenceChar = 0xFE0F;
-
-// emoji modifiers
-static const UTF32Char kType12EmojiModifierChar = 0x1F3FB; // Emoji Modifier Fitzpatrick type-1-2
-static const UTF32Char kType3EmojiModifierChar = 0x1F3FC;  // Emoji Modifier Fitzpatrick type-3
-static const UTF32Char kType4EmojiModifierChar = 0x1F3FD;  // Emoji Modifier Fitzpatrick type-4
-static const UTF32Char kType5EmojiModifierChar = 0x1F3FE;  // Emoji Modifier Fitzpatrick type-5
-static const UTF32Char kType6EmojiModifierChar = 0x1F3FF;  // Emoji Modifier Fitzpatrick type-6
-
+#import "CECharacterInfo.h"
 
 
 @interface CECharacterPopoverController () <NSPopoverDelegate>
 
 @property (nonatomic, nonnull, copy) NSString *glyph;
 @property (nonatomic, nonnull, copy) NSString *unicodeName;
+@property (nonatomic, nullable, copy) NSString *unicodeGroupName;
 @property (nonatomic, nonnull, copy) NSString *unicode;
+
+
+@property (nonatomic, nullable, weak) IBOutlet NSTextField *unicodeGroupNameField;
 
 @end
 
@@ -60,113 +52,37 @@ static const UTF32Char kType6EmojiModifierChar = 0x1F3FF;  // Emoji Modifier Fit
 
 // ------------------------------------------------------
 /// failable initialize
-- (nullable instancetype)initWithCharacter:(nonnull NSString *)character
+- (nullable instancetype)initWithCharacter:(nonnull NSString *)singleString
 // ------------------------------------------------------
 {
-    if ([character numberOfComposedCharacters] != 1) { return nil; }
+    CECharacterInfo *characterInfo = [CECharacterInfo characterInfoWithString:singleString];
+    
+    if (!characterInfo) { return nil; }
     
     self = [super initWithNibName:@"CharacterPopover" bundle:nil];
     if (self) {
-        _glyph = character;
-        
-        NSUInteger length = [character length];
-        
-        // unicode hex
-        NSMutableArray<NSString *> *unicodes = [NSMutableArray array];
-        for (NSUInteger i = 0; i < length; i++) {
-            unichar theChar = [character characterAtIndex:i];
-            unichar nextChar = (length > i + 1) ? [character characterAtIndex:i + 1] : 0;
-            NSString *unicode;
-            
-            if (CFStringIsSurrogateHighCharacter(theChar) && CFStringIsSurrogateLowCharacter(nextChar)) {
-                UTF32Char pair = CFStringGetLongCharacterForSurrogatePair(theChar, nextChar);
-                unicode = [NSString stringWithFormat:@"U+%04tX (U+%04X U+%04X)", pair, theChar, nextChar];
-                i++;
-                
-            } else {
-                unicode = [NSString stringWithFormat:@"U+%04X", theChar];
-            }
-            
-            [unicodes addObject:unicode];
-        }
-        _unicode = [unicodes componentsJoinedByString:@"  "];
-        
-        BOOL isMultipleChars = NO;
-        
-        // check variation selector
-        NSString *variationSelectorAdditional;
-        if ([unicodes count] == 2) {
-            unichar lastChar = [character characterAtIndex:(length - 1)];
-            if (lastChar == kEmojiSequenceChar) {
-                variationSelectorAdditional = @"Emoji Style";
-            } else if (lastChar == kTextSequenceChar) {
-                variationSelectorAdditional = @"Text Style";
-            } else if ((lastChar >= 0x180B && lastChar <= 0x180D) ||
-                       (lastChar >= 0xFE00 && lastChar <= 0xFE0D))
-            {
-                variationSelectorAdditional = @"Variant";
-            } else {
-                unichar highSurrogate = [character characterAtIndex:(length - 2)];
-                unichar lowSurrogate = [character characterAtIndex:(length - 1)];
-                if (CFStringIsSurrogateHighCharacter(highSurrogate) &&
-                    CFStringIsSurrogateLowCharacter(lowSurrogate))
-                {
-                    UTF32Char pair = CFStringGetLongCharacterForSurrogatePair(highSurrogate, lowSurrogate);
-                    
-                    switch (pair) {
-                        case kType12EmojiModifierChar:
-                            variationSelectorAdditional = @"Skin Tone I-II";  // Light Skin Tone
-                            break;
-                        case kType3EmojiModifierChar:
-                            variationSelectorAdditional = @"Skin Tone III";  // Medium Light Skin Tone
-                            break;
-                        case kType4EmojiModifierChar:
-                            variationSelectorAdditional = @"Skin Tone IV";  // Medium Skin Tone
-                            break;
-                        case kType5EmojiModifierChar:
-                            variationSelectorAdditional = @"Skin Tone V";  // Medium Dark Skin Tone
-                            break;
-                        case kType6EmojiModifierChar:
-                            variationSelectorAdditional = @"Skin Tone VI";  // Dark Skin Tone
-                            break;
-                        default:
-                            if (pair >= 0xE0100 && pair <= 0xE01EF) {
-                                variationSelectorAdditional = @"Variant";
-                            } else {
-                                isMultipleChars = YES;
-                            }
-                            break;
-                    }
-                }
-            }
-        } else if ([unicodes count] > 2) {
-            isMultipleChars = YES;
-        }
-        
-        if (isMultipleChars) {
-            // number of characters message
-            _unicodeName = [NSString stringWithFormat:NSLocalizedString(@"<a letter consisting of %d characters>", nil), [unicodes count]];
-            
-        } else {
-            // unicode character name
-            NSMutableString *unicodeName = [character mutableCopy];
-            // You can't use kCFStringTransformToUnicodeName instead of `Any-Name` here,
-            // because some characters (e.g. normal `a`) don't return their name when use this constant.
-            CFStringTransform((__bridge CFMutableStringRef)unicodeName, NULL, CFSTR("Any-Name"), NO);
-            
-            NSRegularExpression *regex = [NSRegularExpression regularExpressionWithPattern:@"\\{(.+?)\\}" options:0 error:nil];
-            NSTextCheckingResult *firstMatch = [regex firstMatchInString:unicodeName options:0
-                                                                   range:NSMakeRange(0, [unicodeName length])];
-            _unicodeName = [unicodeName substringWithRange:[firstMatch rangeAtIndex:1]];
-            
-            if (variationSelectorAdditional) {
-                _unicodeName = [NSString stringWithFormat:@"%@ (%@)", _unicodeName,
-                                NSLocalizedString(variationSelectorAdditional, nil)];
-            }
-        }
+        _glyph = [characterInfo string];
+        _unicode = [[characterInfo unicodes] componentsJoinedByString:@"  "];
+        _unicodeName = [characterInfo unicodeName];
+        _unicodeGroupName = [characterInfo localizedUnicodeGroupName];
     }
     return self;
 }
+
+
+// ------------------------------------------------------
+/// modify xib items
+- (void)viewDidAppear
+// ------------------------------------------------------
+{
+    [super viewDidAppear];
+    
+    // remove group name field if not exists
+    if (![self unicodeGroupName]) {
+        [[self unicodeGroupNameField] removeFromSuperviewWithoutNeedingDisplay];
+    }
+}
+
 
 
 // ------------------------------------------------------